Add functions to create FileChooserBuilder and replace duplicate code with said function
