+++ CoreTesting/src/au/gov/asd/tac/constellation/testing/ExportGlyphTexturesAction.java
@@ -20,9 +20,6 @@ import au.gov.asd.tac.constellation.utilities.gui.filechooser.FileChooser;
 import au.gov.asd.tac.constellation.visual.opengl.utilities.SharedDrawable;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
-import java.io.File;
-import javax.swing.filechooser.FileFilter;
-import org.apache.commons.lang3.StringUtils;
 import org.openide.awt.ActionID;
 import org.openide.awt.ActionReference;
 import org.openide.awt.ActionRegistration;
@@ -54,21 +51,6 @@ public final class ExportGlyphTexturesAction implements ActionListener {
      * @return the created file chooser.
      */
     public FileChooserBuilder getExportGlyphTexturesFileChooser() {
-        return new FileChooserBuilder(TITLE)
-                .setTitle(TITLE)
-                .setAcceptAllFileFilterUsed(false)
-                .setFilesOnly(true)
-                .setFileFilter(new FileFilter() {
-                    @Override
-                    public boolean accept(final File file) {
-                        final String name = file.getName();
-                        return (file.isFile() && StringUtils.endsWithIgnoreCase(name, FileExtensionConstants.PNG)) || file.isDirectory();
-                    }
-
-                    @Override
-                    public String getDescription() {
-                        return "Image Files (" + FileExtensionConstants.PNG + ")";
-                    }
-                });
+        return FileChooser.createFileChooserBuilder(TITLE, FileExtensionConstants.PNG, "Image Files (" + FileExtensionConstants.PNG + ")");
     }
 }
