+++ CoreUtilities/src/au/gov/asd/tac/constellation/utilities/gui/filechooser/FileChooser.java
@@ -24,55 +24,55 @@ import java.util.concurrent.CompletableFuture;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import javafx.application.Platform;
+import javax.swing.JOptionPane;
 import javax.swing.SwingUtilities;
+import javax.swing.filechooser.FileFilter;
+import org.apache.commons.io.FilenameUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.openide.filesystems.FileChooserBuilder;
 
 /**
- * Wrapper for opening file choosers built with the {@link FileChooserBuilder}. There
- * are a couple of issues on macOS that require this to be done in a rather specific
- * way to ensure that the dialog is displayed and the UI does not hang.
+ * Wrapper for opening file choosers built with the {@link FileChooserBuilder}. There are a couple of issues on macOS
+ * that require this to be done in a rather specific way to ensure that the dialog is displayed and the UI does not
+ * hang.
  *
  * @author formalhaunt
  */
 public class FileChooser {
 
     private static final Logger LOGGER = Logger.getLogger(FileChooser.class.getName());
-    
+
     /**
      * Private constructor to prevent external initialization.
      */
     private FileChooser() {
     }
-    
+
     /**
-     * Immediately construct a file chooser using the passed builder and opens it as a save
-     * dialog.
+     * Immediately construct a file chooser using the passed builder and opens it as a save dialog.
      * <p/>
-     * This is intended to be called when we already know we are in the EDT / UI Thread
-     * This ensures correct program flow, with dialogs shown in the correct order,
-     * specifically for a special case where dialogs are cascaded into more dialogs 
-     * and need an immediate result from the new dialogs
-     * 
+     * This is intended to be called when we already know we are in the EDT / UI Thread This ensures correct program
+     * flow, with dialogs shown in the correct order, specifically for a special case where dialogs are cascaded into
+     * more dialogs and need an immediate result from the new dialogs
+     *
      * @param fileChooserBuilder the file chooser builder to build and display
-     * @return a {@link CompletableFuture} that will return the selected file or
-     *     an empty optional if the user selects cancel
+     * @return a {@link CompletableFuture} that will return the selected file or an empty optional if the user selects
+     * cancel
      */
     public static CompletableFuture<Optional<File>> openImmediateSaveDialog(final FileChooserBuilder fileChooserBuilder) {
         return CompletableFuture.completedFuture(openFileDialogAndGetFirstFile(fileChooserBuilder, FileChooserMode.SAVE));
     }
 
     /**
-     * Construct a file chooser using the passed builder and opens it as a save
-     * dialog.
+     * Construct a file chooser using the passed builder and opens it as a save dialog.
      * <p/>
-     * If this was called from a UI thread then this will wrap the request
-     * in a async future call that does a recursive call back to this method. This
-     * ensures that the UI thread does not get into a race condition where the
-     * UI thread becomes blocked.
-     * 
+     * If this was called from a UI thread then this will wrap the request in a async future call that does a recursive
+     * call back to this method. This ensures that the UI thread does not get into a race condition where the UI thread
+     * becomes blocked.
+     *
      * @param fileChooserBuilder the file chooser builder to build and display
-     * @return a {@link CompletableFuture} that will return the selected file or
-     *     an empty optional if the user selects cancel
+     * @return a {@link CompletableFuture} that will return the selected file or an empty optional if the user selects
+     * cancel
      */
     public static CompletableFuture<Optional<File>> openSaveDialog(final FileChooserBuilder fileChooserBuilder) {
         if (SwingUtilities.isEventDispatchThread() || Platform.isFxApplicationThread()) {
@@ -84,19 +84,17 @@ public class FileChooser {
             return CompletableFuture.completedFuture(openFileDialogAndGetFirstFile(fileChooserBuilder, FileChooserMode.SAVE));
         }
     }
-    
+
     /**
-     * Construct a file chooser using the passed builder and opens it as a open
-     * dialog.
+     * Construct a file chooser using the passed builder and opens it as a open dialog.
      * <p/>
-     * If this was called from a UI thread then this will wrap the request
-     * in a async future call that does a recursive call back to this method. This
-     * ensures that the UI thread does not get into a race condition where the
-     * UI thread becomes blocked.
-     * 
+     * If this was called from a UI thread then this will wrap the request in a async future call that does a recursive
+     * call back to this method. This ensures that the UI thread does not get into a race condition where the UI thread
+     * becomes blocked.
+     *
      * @param fileChooserBuilder the file chooser builder to build and display
-     * @return a {@link CompletableFuture} that will return the selected file or
-     *     an empty optional if the user selects cancel
+     * @return a {@link CompletableFuture} that will return the selected file or an empty optional if the user selects
+     * cancel
      */
     public static CompletableFuture<Optional<File>> openOpenDialog(final FileChooserBuilder fileChooserBuilder) {
         if (SwingUtilities.isEventDispatchThread() || Platform.isFxApplicationThread()) {
@@ -108,19 +106,17 @@ public class FileChooser {
             return CompletableFuture.completedFuture(openFileDialogAndGetFirstFile(fileChooserBuilder, FileChooserMode.OPEN));
         }
     }
-    
+
     /**
-     * Construct a file chooser using the passed builder and opens it as a multi
-     * dialog.
+     * Construct a file chooser using the passed builder and opens it as a multi dialog.
      * <p/>
-     * If this was called from a UI thread then this will wrap the request
-     * in a async future call that does a recursive call back to this method. This
-     * ensures that the UI thread does not get into a race condition where the
-     * UI thread becomes blocked.
-     * 
+     * If this was called from a UI thread then this will wrap the request in a async future call that does a recursive
+     * call back to this method. This ensures that the UI thread does not get into a race condition where the UI thread
+     * becomes blocked.
+     *
      * @param fileChooserBuilder the file chooser builder to build and display
-     * @return a {@link CompletableFuture} that will return the selected file(s) or
-     *     an empty optional if the user selects cancel
+     * @return a {@link CompletableFuture} that will return the selected file(s) or an empty optional if the user
+     * selects cancel
      */
     public static CompletableFuture<Optional<List<File>>> openMultiDialog(final FileChooserBuilder fileChooserBuilder) {
         if (SwingUtilities.isEventDispatchThread() || Platform.isFxApplicationThread()) {
@@ -132,20 +128,19 @@ public class FileChooser {
             return CompletableFuture.completedFuture(openFileDialog(fileChooserBuilder, FileChooserMode.MULTI));
         }
     }
-    
+
     /**
-     * Construct a file chooser using the passed builder and opens it as a open
-     * or save dialog depending on the passed mode. This method will block the
-     * calling thread until the user selects a file or presses cancel.
+     * Construct a file chooser using the passed builder and opens it as a open or save dialog depending on the passed
+     * mode. This method will block the calling thread until the user selects a file or presses cancel.
      *
      * @param fileChooserBuilder the file chooser builder to build and display
      * @param fileDialogMode the type of file chooser dialog to open, save, open or multi
      * @return the selected file(s) or an empty optional if the user selects cancel
      */
-    private static Optional<List<File>> openFileDialog(final FileChooserBuilder fileChooserBuilder, 
+    private static Optional<List<File>> openFileDialog(final FileChooserBuilder fileChooserBuilder,
             final FileChooserMode fileDialogMode) {
         final ShowFileChooserDialog showDialog = new ShowFileChooserDialog(fileChooserBuilder, fileDialogMode);
-        
+
         // Check if the calling thread is able to run this
         if (SwingUtilities.isEventDispatchThread() || Platform.isFxApplicationThread()) {
             showDialog.run();
@@ -163,23 +158,65 @@ public class FileChooser {
 
         return showDialog.getSelectedFiles();
     }
-    
+
     /**
-     * Opens a file dialog with the selected mode and returns the first selected file
-     * or an empty optional if the user selected cancel.
+     * Opens a file dialog with the selected mode and returns the first selected file or an empty optional if the user
+     * selected cancel.
      *
      * @param fileChooserBuilder the file chooser builder to build and display
      * @param fileDialogMode the type of file chooser dialog to open, save, open or multi
      * @return the first selected file or an empty optional if the user selects cancel
-     * @see #openFileDialog(FileChooserBuilder, FileChooserMode) 
+     * @see #openFileDialog(FileChooserBuilder, FileChooserMode)
      */
-    private static Optional<File> openFileDialogAndGetFirstFile(final FileChooserBuilder fileChooserBuilder, 
+    private static Optional<File> openFileDialogAndGetFirstFile(final FileChooserBuilder fileChooserBuilder,
             final FileChooserMode fileDialogMode) {
         final Optional<List<File>> selectedFiles = openFileDialog(fileChooserBuilder, fileDialogMode);
-        
+
         if (selectedFiles.isPresent() && !selectedFiles.get().isEmpty()) {
             return Optional.of(selectedFiles.get().get(0));
         }
         return Optional.empty();
+    }
+
+    private static boolean approver(final File[] selection, final String fileExtension) {
+
+        // Show dialog box if file already exists when saving
+        final String extension = FilenameUtils.getExtension(selection[0].toString());
+        final String filepath = extension.isEmpty() ? selection[0].toString() + fileExtension : selection[0].toString();
+        final File file = new File(filepath);
+
+        if (file.exists()) {
+            int response = JOptionPane.showConfirmDialog(null,
+                    "The file " + file.getName() + " already exists. Do you want to replace the existing file?",
+                    "Overwrite file", JOptionPane.YES_NO_OPTION, JOptionPane.WARNING_MESSAGE);
+            // Overwrite if user chose yes
+            return response == JOptionPane.YES_OPTION;
+        }
+        // Otherwise file didn't already exist
+        return true;
+    }
+    
+    public static FileChooserBuilder createFileChooserBuilderNoFilter(final String title, final String fileExtension) {
+        return new FileChooserBuilder(title)
+                .setTitle(title)
+                .setFilesOnly(true)
+                .setSelectionApprover((File[] selection) -> approver(selection, fileExtension));
+    }
+    
+    public static FileChooserBuilder createFileChooserBuilder(final String title, final String fileExtension, final String description) {
+        return createFileChooserBuilderNoFilter(title, fileExtension)
+                .setAcceptAllFileFilterUsed(false)
+                .setFileFilter(new FileFilter() {
+                    @Override
+                    public boolean accept(final File file) {
+                        final String name = file.getName();
+                        return (file.isFile() && StringUtils.endsWithIgnoreCase(name, fileExtension)) || file.isDirectory();
+                    }
+
+                    @Override
+                    public String getDescription() {
+                        return description;
+                    }
+                });
     }
 }
